+++ src/TokenGenerator.php
@@ -9,13 +9,20 @@ class TokenGenerator
     private string $csrfToken;
 
     /**
-     * Generates a CSRF token and saves it to the session.
+     * Generates a CSRF token.
+     * Saves the token to the session if the user is logged in.
+     * @param bool $loggedUser Indicates if the user is logged in. Default is true.
+     * 
      * @return string The generated CSRF token.
      */
-    public function generate(): string
+    public function generate(bool $loggedUser = true): string
     {
         $this->csrfToken = bin2hex(random_bytes(32));
-        $_SESSION['csrf_token'] = $this->csrfToken;
+
+        if ($loggedUser) {
+            $_SESSION['csrf_token'] = $this->csrfToken;
+        }
+
         return $this->csrfToken;
     }
 }